Revert "Issue #2809533 by Lendude, jibran: Convert AJAX part of \Drupal\system\Tests\Ajax\FrameworkTest to JavascriptTestBase"
This reverts commit 3c7cf83edfc61a19faf867b3ad3c9e0529c609ef.
